requirements.txt
transformers[torch]==4.37.2
flash-attn==2.7.2.post1
fastapi==0.115.6
pillow==11.1.0
colorlog==6.9.0
uvicorn==0.34.0
python-multipart==0.0.20
pyngrok==7.2.2
ultralytics==8.3.56